# specs/001-phase-7-intelligence/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Physical AI Curriculum API - Phase 7 Final Intelligence
  description: |
    API for adaptive content personalization and multi-language support.
    
    ## Features
    - User profile management with hardware configuration and skill level
    - Personalized chapter summaries based on hardware profile and skill level
    - Roman Urdu translation toggle for chapter content
    - Complete 13-week curriculum access
    
    ## Authentication
    All endpoints require JWT Bearer token except where noted.
    
    ## Hardware Profiles
    - `sim_rig`: Desktop setup with RTX 4070 Ti or better
    - `edge_kit`: Jetson Orin Nano/NX or AGX Orin
    - `unitree`: Unitree Go2, G1, or Proxy robots
    
    ## Skill Levels
    - `beginner`: Foundational concepts, simple language
    - `intermediate`: Technical details, assumes basic knowledge
    - `advanced`: Deep optimization, production considerations
  version: 1.0.0
  contact:
    name: Physical AI Team
servers:
  - url: https://api.physicalai.example.com/api/v1
    description: Production server
  - url: http://localhost:8000/api/v1
    description: Local development server
...
paths:
...
  /chapters/{chapter_id}/summary:
    get:
      tags:
        - Personalization
      summary: Get personalized chapter summary
      description: |
        Retrieve chapter summary personalized for the user's hardware profile and skill level.
        
        If no query parameters provided, uses current user's profile.
        Admins can request summaries for specific hardware/skill combinations.
      operationId: getChapterSummary
      security:
        - bearerAuth: []
      parameters:
        - name: chapter_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: hardware_profile
          in: query
          required: false
          schema:
            type: string
            enum: [sim_rig, edge_kit, unitree]
          description: Override hardware profile (admin only)
        - name: skill_level
          in: query
          required: false
          schema:
            type: string
            enum: [beginner, intermediate, advanced]
          description: Override skill level (admin only)
      responses:
        '200':
          description: Personalized summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChapterSummaryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Chapter not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /personalization/regenerate:
    post:
      tags:
        - Personalization
      summary: Regenerate all summaries
      description: |
        Regenerate personalized summaries for all chapters.
        
        **Admin only**. Useful after content updates or prompt improvements.
      operationId: regenerateSummaries
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                hardware_profile:
                  type: string
                  enum: [sim_rig, edge_kit, unitree]
                skill_level:
                  type: string
                  enum: [beginner, intermediate, advanced]
              description: Filter by specific combination (optional)
      responses:
        '202':
          description: Regeneration started (background job)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /chapters/{chapter_id}/translation:
    get:
      tags:
        - Translations
      summary: Get chapter translation
      description: |
        Retrieve Roman Urdu translation of chapter content.
        
        Returns 404 if translation not available (show English with indicator per spec).
      operationId: getChapterTranslation
      security:
        - bearerAuth: []
      parameters:
        - name: chapter_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: lang
          in: query
          required: true
          schema:
            type: string
            example: ur-Latn
          description: Language code (ur-Latn for Roman Urdu)
      responses:
        '200':
          description: Translation found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Translation not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags:
        - Translations
      summary: Update translation
      description: |
        Update or create translation for a chapter.
        
        **Admin only**. Used for manual translation updates or AI translation review.
      operationId: updateTranslation
      security:
        - bearerAuth: []
      parameters:
        - name: chapter_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslationUpdate'
      responses:
        '200':
          description: Successful update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /translations/status:
    get:
      tags:
        - Translations
      summary: Get translation coverage stats
      description: |
        Retrieve translation progress statistics.
        
        Shows how many chapters have been translated to Roman Urdu.
      operationId: getTranslationStatus
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Translation statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationStatsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication endpoint
  schemas:
...
    ChapterSummaryResponse:
      type: object
      properties:
        chapter_id:
          type: string
          format: uuid
        summary_content:
          type: string
          description: Markdown-formatted personalized summary
        hardware_profile_type:
          type: string
          enum: [sim_rig, edge_kit, unitree]
        skill_level:
          type: string
          enum: [beginner, intermediate, advanced]
        generated_at:
          type: string
          format: date-time
    
    # Translation Schemas
    TranslationResponse:
      type: object
      properties:
        chapter_id:
          type: string
          format: uuid
        translated_content:
          type: string
          description: Markdown-formatted Roman Urdu translation
        language_code:
          type: string
          example: ur-Latn
        status:
          type: string
          enum: [draft, in_review, published]
        updated_at:
          type: string
          format: date-time
    TranslationUpdate:
      type: object
      required:
        - language_code
        - translated_content
      properties:
        language_code:
          type: string
          example: ur-Latn
        translated_content:
          type: string
        status:
          type: string
          enum: [draft, in_review, published]
          default: draft
        review_notes:
          type: string
    TranslationStatsResponse:
      type: object
      properties:
        total_chapters:
          type: integer
        translated_chapters:
          type: integer
        published_chapters:
          type: integer
        draft_chapters:
          type: integer
        coverage_percentage:
          type: number
          format: float
        by_language:
          type: object
          additionalProperties:
            type: object
            properties:
              published:
                type: integer
              draft:
                type: integer
              in_review:
                type: integer
    
    # Curriculum Schemas
...
    JobStatusResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed]
        message:
          type: string
        created_at:
          type: string
          format: date-time
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          description: Additional error details (validation errors, etc.)
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "unauthorized"
            message: "Authentication required"